Skip making a PID namespace if /proc cannot be mounted in fakeroot build

+++ internal/pkg/runtime/engine/fakeroot/engine_linux.go
@@ -104,7 +104,20 @@ func (e *EngineOperations) PrepareConfig(starterConfig *starter.Config) error {
 
 	g.AddOrReplaceLinuxNamespace(specs.UserNamespace, "")
 	g.AddOrReplaceLinuxNamespace(specs.MountNamespace, "")
-	g.AddOrReplaceLinuxNamespace(specs.PIDNamespace, "")
+
+	// Check if there are bind mounts under /proc that would prevent
+	// a new /proc mount from being created in a PID namespace
+	hasMounts, err := proc.HasBindMountsUnderProc()
+	if err == nil && hasMounts {
+		sylog.Infof("Cannot mount /proc, proceeding without PID namespace")
+		e.EngineConfig.HasPIDNamespace = false
+	} else {
+		if err != nil {
+			sylog.Debugf("error checking /proc mounts, proceeding: %s", err)
+		}
+		g.AddOrReplaceLinuxNamespace(specs.PIDNamespace, "")
+		e.EngineConfig.HasPIDNamespace = true
+	}
 
 	uid, err := safecast.Convert[uint32](os.Getuid())
 	if err != nil {
@@ -307,25 +320,28 @@ func (e *EngineOperations) StartProcess(_ int) error {
 	if err != nil {
 		return fmt.Errorf("failed to mount %s to /root: %s", e.EngineConfig.Home, err)
 	}
-	tmpdir, err := os.MkdirTemp("", "bind-mount-*")
-	if err != nil {
-		return fmt.Errorf("failed to create temporary directory: %s", err)
-	}
-	err = syscall.Mount("proc", tmpdir, "proc", syscall.MS_NOSUID|syscall.MS_NOEXEC|syscall.MS_NODEV, "")
-	if err != nil {
-		return fmt.Errorf("failed to mount proc filesystem: %s", err)
-	}
-	err = syscall.Mount(binfmtMisc, filepath.Join(tmpdir, "sys", "fs", "binfmt_misc"), "", syscall.MS_BIND, "")
-	if err != nil {
-		return fmt.Errorf("failed to mount binfmt_misc filesystem: %s", err)
-	}
-	err = syscall.Mount(tmpdir, "/proc", "", syscall.MS_MOVE, "")
-	if err != nil {
-		return fmt.Errorf("failed to mount proc filesystem: %s", err)
-	}
-	err = os.Remove(tmpdir)
-	if err != nil {
-		return fmt.Errorf("failed to remove temporary directory: %s", err)
+
+	if e.EngineConfig.HasPIDNamespace {
+		tmpdir, err := os.MkdirTemp("", "bind-mount-*")
+		if err != nil {
+			return fmt.Errorf("failed to create temporary directory: %s", err)
+		}
+		err = syscall.Mount("proc", tmpdir, "proc", syscall.MS_NOSUID|syscall.MS_NOEXEC|syscall.MS_NODEV, "")
+		if err != nil {
+			return fmt.Errorf("failed to mount proc filesystem: %s", err)
+		}
+		err = syscall.Mount(binfmtMisc, filepath.Join(tmpdir, "sys", "fs", "binfmt_misc"), "", syscall.MS_BIND, "")
+		if err != nil {
+			return fmt.Errorf("failed to mount binfmt_misc filesystem: %s", err)
+		}
+		err = syscall.Mount(tmpdir, "/proc", "", syscall.MS_MOVE, "")
+		if err != nil {
+			return fmt.Errorf("failed to mount proc filesystem: %s", err)
+		}
+		err = os.Remove(tmpdir)
+		if err != nil {
+			return fmt.Errorf("failed to remove temporary directory: %s", err)
+		}
 	}
 
 	// fix potential issue with SELinux (https://github.com/apptainer/singularity/issues/4038)
